add eliot fluentd logging capability to subscription manager
